back/api/requirements.txt
Django>=5.0,<6
djangorestframework>=3.16
psycopg2-binary>=2.9
django-cors-headers>=4.4
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
